Fetch bugs in batches
User can now defined a start date for Bugzilla from where to start
fetching bugs, and fetch_days to define how many days from start_date to
end_date to include in each query. This way we do multiple requests to
Bugzilla but fetch significantly less data on a single query when
comparing to not defining these and having to fetch all bugs at once. If
there is a huge amount of bugs the request will likely time out and all
will fail.