QA Reports base URL from config file
Generate the URL to reports on server side when performing DB queries to
get rid of hard-coded URLs in client widgets.

The reason for creating the URL in QA Dashboard instead of receiving it
from QA Reports was done mainly because of QA Reports exports single
reports when created/updated using an Observer and a model, neither of
which have access to url_for method if hostname is not provided.

We do anyway support only one QA Reports server so this will work just
fine for now at least.