Use Bugzilla service prefixes
QA Reports nowadays supports multiple Bugzilla services. Each service
has a unique prefix with which the correct service can be identified. It
also returns the bugs for each test case as a list of objects containing
ID, prefix, and URL from the endpoints from where QA Dashboard gets it's
data.

So, store the bugs with the prefix in place to the database (single
unique key still works). Then when fetching bug counts get other data
including the bug URL as well and not just the bug count. This way we
don't need hardcoded URLs or such.

This multi-Bugzilla approach now works for the QA Dashboard application.
Need to change Bugzilla exporter still to support multiple Bugzilla
servers.

It is also worth to notice that at this point old QA Reports
installations are not anymore compliant with QA Dashboard and vice
versa.