Flexible filtering implemented
Now filtering options in left column in edit mode do not restrict the
values that can be selected offering greater flexibility in creating
widgets.

Now user can, if he wishes, show reports from e.g. different releases in
the same widget (earlier only one widget could be selected). Due to this
the title for each item in the widget now includes the whole
release/profile/testset/product combo because otherwise it would be
impossible to know what the partial part of the widget is showing.

Things to do still include:
- Widget title
- Default config