Moved create radio button to WidgetBase
QAReportsWidget had method for generating radio buttons. It did not
require access to any class instance data or such, and it is useful for
other widgets deriving from WidgetBase, so moved the function there.

In general the function is more of a helper method and should probably
be placed completely outside the widgets but leaving it there for now.