QA Reports widget data filtering and API change
The radio buttons in the left column of edit view now only filter the
suggestions list when earlier they filtered also selected sets lists
and, if user had selected sets not matching the filters when closing
edit mode, also narrowed down the sets actually shown in the widget.

Now the filter buttons only filter the suggestions list so user can use
them to narrow down the list of suggestions but build the selected sets
how he wishes.

The API was changed as well so that it accepts the groups as HTTP POST
parameters instead of just the filter values meaning we now fetch only
the selected sets. Earlier if user had collected sets from here and
there and then set the filters to "Any" to get the data actually visible
we requested way more data than needed from the backend.