Convert tlsOptions.ca
tls module wants the ca-option (amongs others) to be an array of
buffers. Since the LDAP settings are defined in a JSON file we kind of
expect it to be a file name.

So, if the file is found convert it. If not then do nothing and probably
fail. There are other possible options that should be handled the same
way but in current environment I just need the ca certificate.